Created definitions.h to ease header includes
